WIT mode in EPEC::Game::iterativeNash
WIT (Whatever It Takes) is triggered when the user requires a pure NE . The algorithm first finds a MNE (if exists). When the former is found, the algorithm triggers the WIT mode (if the eq is Mixed and not Pure), and tries to check whether a pure equilibrium exists among the currently enumerated polyhedra. If not, it randomly add polyhedra untill all are enumerated (or timelimit triggers). Hence, it finds a pure NE, it will return it. Otherwise, it will return the MNE minimizing the number of polyhedra emloyed.